MLDB-1163 split text loading from dataset to procedure
