+++ testing/MLDB-974-slow-subquery.js
@@ -1,14 +1,17 @@
 // This file is part of MLDB. Copyright 2015 Datacratic. All rights reserved.
 
-var dsconfig = {
-    id: 'test',
-    type: "text.csv.tabular",
+csv_conf = {
+    type: "import.text",
     params: {
-        dataFileUrl: "https://s3.amazonaws.com/benchm-ml--main/test.csv"
+        dataFileUrl : "https://s3.amazonaws.com/benchm-ml--main/test.csv",
+        ouputDataset: {
+            id: "test",
+        },
+        runOnCreation: true,     
     }
-};
+}
 
-var dataset = mldb.createDataset(dsconfig);
+var res = mldb.put("/v1/procedures/csv_proc", csv_conf)
 
 var before = new Date();
